refactor: extract _resolve_steam_id to BaseEndpoint
Moves the shared _resolve_steam_id method from individual endpoint
classes to BaseEndpoint, reducing code duplication across 5 files.
